All algorithms can now work with APCGs that contain multiple tasks per core. The method which prints the mapping was also improved.

+++ src/ro/ulbsibiu/acaps/mapper/bb/ProcComm.java
@@ -19,7 +19,7 @@ class ProcComm {
 	public int dstProc;
 
 	/** the bandwidth requirement for this communication */
-	public int bandwidth;
+	public long bandwidth;
 
 	/** (only useful in routing synthesis) */
 	public int adaptivity;
